requirements.txt
open_clip_torch
gradio
numpy
torch
pytorch_lightning
